requirements.txt
alembic==1.11.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pipreqs==0.4.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
WTForms==3.0.1
yarg==0.1.9
